+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add sort to exploration preview component [#20](https://github.com/opendatateam/udata-tabular-preview/pull/20)
 - Fix setuptools version used in CI [#23](https://github.com/opendatateam/udata-tabular-preview/pull/23)
 - Add pagination to exploration preview component [#21](https://github.com/opendatateam/udata-tabular-preview/pull/21)
+- Add data structure component and explore button for hook `dataset.display.explore-button` [#19](https://github.com/opendatateam/udata-tabular-preview/pull/19)
 
 ## 3.0.0 (2023-01-18)